Fixed frame number issue
When the grab() function is called, the frame is not guaranteed to
be written to the output svo file. This means you might increment frame
number, where it shouldn't be incremented (frame wasn't written).
Thus, we need to track get_recording_status().status parameter.

Some commented parts were removed as it is not relevant anymore.