Script re-written completely, solved all issues
Previously, was using numpy.dumps and numpy.loads, which seem depricated
Now using pickle to convert the data into bytes

Also, previously was sending truncated data, now implemented a system
that will loop over the length of the data to receive it completely.
The system first receives the length of the data in bytes, then the
data itself.